aktueller Stand, der Auslagerung von GameState.cs
